pinned-requirements.txt
# This file was autogenerated via `uv export`.
-e .
aiostream==0.6.2
ansicon==1.89.0 ; platform_system == 'Windows'
anyio==4.4.0
apischema==0.18.1
asttokens==2.4.1
attrs==24.2.0
authlib==1.3.2
backoff==2.2.1
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blessed==1.20.0
cattrs==24.1.1
certifi==2024.8.30
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2
click==8.1.7
codefind==0.1.7
coleo==0.3.4
colorama==0.4.6 ; sys_platform == 'win32' or platform_system == 'Windows'
coverage==7.6.1
cryptography==43.0.1
deprecated==1.2.14
dnspython==2.6.1
eventlet==0.37.0
exceptiongroup==1.2.2 ; python_full_version < '3.11'
executing==2.1.0
fake-useragent==1.5.1
future==1.0.0
gifnoc==0.4.0
giving==0.4.3
greenlet==3.1.0
h11==0.14.0
hrepr==0.8.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinxed==1.3.0 ; platform_system == 'Windows'
jurigged==0.6.0
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
openreview-py==1.43.2
ovld==0.4.2
packaging==24.1
platformdirs==4.3.3
pluggy==1.5.0
prompt-toolkit==3.0.36
ptera==1.4.1
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pycryptodome==3.20.0
pydantic==1.10.18
pygments==2.18.0
pyjwt==2.9.0
pylatexenc==2.10
pyparsing==3.1.4
pytest==8.3.3
pytest-cov==5.0.0
pytest-datadir==1.5.0
pytest-regressions==2.5.0
python-multipart==0.0.9
pyyaml==6.0.2
questionary==2.0.1
reactivex==4.0.4
requests==2.32.3
requests-cache==1.2.1
rich==13.8.1
sentry-sdk==2.14.0
setuptools==65.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==1.4.54
sse-starlette==2.1.3
starbear==0.2.2
starlette==0.38.5
tld==0.13
tomli==2.0.1 ; python_full_version <= '3.11'
tqdm==4.66.5
typing-extensions==4.12.2
unidecode==1.3.8
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.30.6
varname==0.13.3
watchdog==5.0.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.0.1
wrapt==1.16.0
